Move requests from /common to /requests. Move semantic and easy to find. Also, change list item component file names to have 'ListItem' in the filename.
