+++ src/components/common/index.js
@@ -1,7 +1,3 @@
 export * from './Input';
 export * from './Button';
 export * from './Spinner';
-export * from './Request';
-export * from './NonRequest';
-export * from './ReceivedRequest';
-
