+++ src/components/requests/NonRequest.js
@@ -24,4 +24,4 @@ const styles = StyleSheet.create({
   }
 });
 
-export { NonRequest };
+export default NonRequest;
