+++ src/components/requests/SendRequest.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { connect } from 'react-redux';
 import Firebase from 'firebase';
-import { Request } from '../common';
+import Request from './Request';
 import { changeRequestStatus } from '../../actions';
 
 class SendRequest extends React.Component {
